+++ src/engine/Graphics/Font.ts
@@ -272,6 +272,36 @@ export class Font extends Graphic implements FontRenderer {
     return ctx;
   }
 
+  private _splitTextBitmap(bitmap: CanvasRenderingContext2D) {
+    const textImages: {x: number, y: number, canvas: HTMLCanvasElement}[] = [];
+    let currentX = 0;
+    let currentY = 0;
+    // 4k is the max for mobile devices
+    const width = Math.min(4096, bitmap.canvas.width);
+    const height = Math.min(4096, bitmap.canvas.height);
+
+    // Splits the original bitmap into 4k max chunks
+    while (currentX < bitmap.canvas.width) {
+      while (currentY < bitmap.canvas.height) {
+        // create new bitmap
+        const canvas = document.createElement('canvas');
+        canvas.width = width;
+        canvas.height = height;
+        const ctx = canvas.getContext('2d');
+
+        // draw current slice to new bitmap in < 4k chunks
+        ctx.drawImage(bitmap.canvas, currentX, currentY, width, height, 0, 0, width, height);
+
+        textImages.push({x: currentX, y: currentY, canvas});
+        currentY += height;
+      }
+      currentX += width;
+      currentY = 0;
+    }
+    return textImages;
+  }
+
+  private _textFragments: {x: number, y: number, canvas: HTMLCanvasElement}[] = [];
   public render(ex: ExcaliburGraphicsContext, text: string, colorOverride: Color, x: number, y: number) {
     if (this.showDebug) {
       this.clearCache();
@@ -284,8 +314,6 @@ export class Font extends Graphic implements FontRenderer {
     // Bounds of the text
     this._textBounds = this.measureText(text);
 
-    // TODO if the text is bigger than 4k we need to split it somehow
-
     if (isNewBitmap) {
       // Setting dimension is expensive because it invalidates the bitmap
       this._setDimension(this._textBounds, bitmap);
@@ -297,29 +325,36 @@ export class Font extends Graphic implements FontRenderer {
     const lines = text.split('\n');
     const lineHeight = this._textBounds.height / lines.length;
 
-
-    const rasterWidth = bitmap.canvas.width;
-    const rasterHeight = bitmap.canvas.height;
-
     if (isNewBitmap) {
       // draws the text to the bitmap
       this._drawText(bitmap, text, colorOverride, lineHeight);
-      // draws the bitmap to excalibur graphics context
-      TextureLoader.load(bitmap.canvas, this.filtering, true);
+
+      // clean up any existing fragments
+      for (const frag of this._textFragments) {
+        TextureLoader.delete(frag.canvas);
+      }
+
+      this._textFragments = this._splitTextBitmap(bitmap);
+
+      for (const frag of this._textFragments) {
+        TextureLoader.load(frag.canvas, this.filtering, true);
+      }
     }
 
-    // Send draw cal to the ex graphics context
-    ex.drawImage(
-      bitmap.canvas,
-      0,
-      0,
-      rasterWidth,
-      rasterHeight,
-      x - rasterWidth / this.quality / 2,
-      y - rasterHeight / this.quality / 2,
-      rasterWidth / this.quality,
-      rasterHeight / this.quality
-    );
+    // draws the bitmap fragments to excalibur graphics context
+    for (const frag of this._textFragments) {
+      ex.drawImage(
+        frag.canvas,
+        0,
+        0,
+        frag.canvas.width,
+        frag.canvas.height,
+        frag.x / this.quality + x - bitmap.canvas.width / this.quality / 2,
+        frag.y / this.quality + y - bitmap.canvas.height / this.quality / 2,
+        frag.canvas.width / this.quality,
+        frag.canvas.height / this.quality
+      );
+    }
 
     this._postDraw(ex);
 
@@ -350,6 +385,7 @@ export class Font extends Graphic implements FontRenderer {
       // if bitmap hasn't been used in 1 second clear it
       if (time + 1000 < performance.now()) {
         this._bitmapUsage.delete(bitmap);
+        TextureLoader.delete(bitmap.canvas);
       }
     }
   }